requirements.txt
pandas
mysql-connector-python
streamlit
folium
